Warning fixes for 64bit visual c++ on Windows

+++ Source/CParse/templ.c
@@ -237,7 +237,7 @@ String *partial_arg(String *s, String *p) {
   if (!c) {
     return Copy(s);
   }
-  prefix = NewStringWithSize(cp, c - cp);
+  prefix = NewStringWithSize(cp, (int)(c - cp));
   newarg = Copy(s);
   Replace(newarg, prefix, "", DOH_REPLACE_ANY | DOH_REPLACE_FIRST);
   Delete(prefix);
