+++ Source/Swig/include.c
@@ -194,7 +194,7 @@ static FILE *Swig_open_file(const_String_or_char_ptr name, int sysfile, int use_
     lastpath = filename;
 
     /* Skip the UTF-8 BOM if it's present */
-    nbytes = fread(bom, 1, 3, f);
+    nbytes = (int)fread(bom, 1, 3, f);
     if (nbytes == 3 && bom[0] == (char)0xEF && bom[1] == (char)0xBB && bom[2] == (char)0xBF) {
       /* skip */
     } else {
@@ -369,7 +369,7 @@ String *Swig_file_filename(const_String_or_char_ptr filename) {
 String *Swig_file_dirname(const_String_or_char_ptr filename) {
   const char *delim = SWIG_FILE_DELIMITER;
   const char *c = strrchr(Char(filename), *delim);
-  return c ? NewStringWithSize(filename, c - Char(filename) + 1) : NewString("");
+  return c ? NewStringWithSize(filename, (int)(c - Char(filename) + 1)) : NewString("");
 }
 
 /*
